+++ cocos/audio/android/jni/cddandroidAndroidJavaEngine.h
@@ -33,7 +33,7 @@ namespace CocosDenshion {
             void unloadEffect(const char* pszFilePath);
 
             bool hasVibrator();
-            void vibrate(long time);
+            void vibrate(long long time);
             void cancelVibrate();
 
         private :
