+++ cocos/audio/android/opensl/cddandroidOpenSLEngine.h
@@ -33,7 +33,7 @@ namespace CocosDenshion {
             void preloadEffect(const char* pszFilePath);
             void unloadEffect(const char* pszFilePath);
             bool hasVibrator();
-            void vibrate(long time);
+            void vibrate(long long time);
             void cancelVibrate();
         };
     }
